Tied loose ends for basic execution
Taken GPU and MMU out of Z80.
Added GameBoy class, which includes Z80, MMU and GPU inside.
It also has the dispatch loop thread running.

Added support for all memory regions in MMU.
Added ROM loading function in MMU.

+++ GBEmu/GBEmu/Include/GPU.h
@@ -50,7 +50,7 @@ private:
 	std::map<TileMap::Number, TileMap> tilemaps_;
 	Renderer renderer_;
 	std::vector<uint8_t> framebuffer_;
-	std::map<uint8_t, std::array<uint8_t, 3>> palette_;
+	std::map<size_t, std::array<uint8_t, 3>> palette_;
 	MMU &mmu_;
 
 	Mode current_mode_{ Mode::ReadingOAM };
